+++ src/services/index.ts
@@ -24,6 +24,7 @@ import Lambda from './lambda'
 import NATGateway from './natGateway'
 import NetworkInterface from './networkInterface'
 import VPC from './vpc'
+import SQS from './sqs'
 import APIGatewayRestApi from './apiGatewayRestApi'
 import APIGatewayResource from './apiGatewayResource'
 import APIGatewayStage from './apiGatewayStage'
@@ -56,6 +57,7 @@ export const serviceMap = {
   [services.sg]: AwsSecurityGroup,
   // [services.subnet]: AwsSubnet, // TODO: Enable when going for ENG-222
   [services.vpc]: VPC,
+  [services.sqs]: SQS,
   tag: AwsTag,
 }
 
@@ -287,7 +289,7 @@ export default class Provider extends CloudGraph.Client {
   }
 
   /**
-   * getData is used to fetch all provider data specificed in the config for the provider
+   * getData is used to fetch all provider data specified in the config for the provider
    * @param opts: A set of optional values to configure how getData works
    * @returns Promise<any> All provider data
    */
@@ -378,12 +380,12 @@ export default class Provider extends CloudGraph.Client {
             }
           }
         })
-        result.entities.push({
-          name: serviceData.name,
-          mutation: serviceClass.mutation,
-          data: entities,
-        })
       }
+      result.entities.push({
+        name: serviceData.name,
+        mutation: serviceClass.mutation,
+        data: entities,
+      })
     }
     return result
   }